Refactor still-useful helper to its own file
