+++ src/module/vue/components/with-rolllisteners.vue
@@ -7,7 +7,7 @@
 <script setup lang="ts">
 import { inject, onMounted, ref, useAttrs } from 'vue'
 import { IronswornActor } from '../../actor/actor'
-import { attachInlineRollListeners } from '../../helpers/rolldialog'
+import { attachInlineRollListeners } from '../../features/rollplus'
 import { IronswornItem } from '../../item/item'
 import { $ActorKey } from '../provisions'
